refactoring: Move logic from the midhandler to a separate CaseWorkerService
