+++ src/main/java/uk/gov/hmcts/reform/iacaseapi/domain/entities/roleassignment/Jurisdiction.java
@@ -1,5 +1,5 @@
 package uk.gov.hmcts.reform.iacaseapi.domain.entities.roleassignment;
 
-public enum RoleType {
-    CASE, ORGANISATION
+public enum Jurisdiction {
+    IA
 }
