+++ src/main/java/uk/gov/hmcts/reform/iacaseapi/domain/entities/roleassignment/RoleCategory.java
@@ -1,5 +1,5 @@
 package uk.gov.hmcts.reform.iacaseapi.domain.entities.roleassignment;
 
 public enum RoleCategory {
-    JUDICIAL, STAFF;
+    JUDICIAL, STAFF
 }
